OpenStack-Rabbit-Consumer/requirements.txt
rabbitpy
requests
requests_kerberos
pika
urllib3
openstacksdk